requirements.txt
streamlit
Pillow>=7.1.0
streamlit-cropper
torch
